# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "description": "Yeni kullanıcı kaydı oluşturur. Varsayılan rol: STUDENT",
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kullanıcı başarıyla oluşturuldu"
          },
          "400": {
            "description": "Hatalı istek - Geçersiz e-posta formatı veya eksik alanlar"
          },
          "403": {
            "description": "E-posta zaten kullanımda"
          }
        },
        "summary": "Yeni kullanıcı oluştur",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "description": "E-posta ve şifre ile kullanıcı girişi yapar. Başarılı girişte access ve refresh token döner.",
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kullanıcı başarıyla giriş yaptı - Access ve refresh token döndürülür",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Hatalı istek - Eksik e-posta veya şifre"
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri - Yanlış e-posta veya şifre"
          }
        },
        "summary": "Kullanıcı girişi",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Kullanıcının refresh token'ını geçersiz kılar ve çıkış yapar.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Kullanıcı başarıyla çıkış yaptı"
          },
          "401": {
            "description": "Kimlik doğrulama gerekli - Bearer token eksik veya geçersiz"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcı çıkışı",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Refresh token kullanarak yeni access token alır. Body'de refresh token gönderilmelidir.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token başarıyla yenilendi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz refresh token"
          }
        },
        "summary": "Erişim token'ını yenile",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Giriş yapmış kullanıcının detaylı profil bilgilerini döndürür.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Kullanıcı bilgileri döndürüldü",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Kimlik doğrulama gerekli - Bearer token eksik veya geçersiz"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mevcut kullanıcıyı getir",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired verification token"
          }
        },
        "summary": "Verify email address",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent"
          },
          "400": {
            "description": "Email already verified"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Resend email verification",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Şifre başarıyla değiştirildi"
          },
          "400": {
            "description": "Geçersiz şifre"
          },
          "401": {
            "description": "Yetkisiz erişim"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Şifre değiştir",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Şifre sıfırlama e-postası gönderildi"
          },
          "400": {
            "description": "Geçersiz e-posta"
          }
        },
        "summary": "Şifre sıfırlama talebi",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Şifre başarıyla sıfırlandı"
          },
          "400": {
            "description": "Geçersiz veya süresi dolmuş token"
          }
        },
        "summary": "Şifre sıfırla",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/upload-profile-picture": {
      "post": {
        "description": "Kullanıcının profil fotoğrafını yükler. Desteklenen formatlar: JPG, JPEG, PNG, GIF. Maksimum boyut: 5MB",
        "operationId": "AuthController_uploadProfilePicture",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil fotoğrafı başarıyla yüklendi"
          },
          "400": {
            "description": "Geçersiz dosya formatı veya boyutu"
          },
          "401": {
            "description": "Kimlik doğrulama gerekli"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil fotoğrafı yükle",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/ai-chat/send-message": {
      "post": {
        "operationId": "AiChatController_sendMessage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "AI yanıtı başarıyla alındı"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının AI ile sohbet etmesi için mesaj gönderme",
        "tags": [
          "AI Chat"
        ]
      }
    },
    "/ai-chat/conversation-history": {
      "get": {
        "operationId": "AiChatController_getConversationHistory",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sohbet geçmişi başarıyla alındı"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının sohbet geçmişini getirme",
        "tags": [
          "AI Chat"
        ]
      }
    },
    "/ai-chat/get-study-suggestions": {
      "post": {
        "operationId": "AiChatController_getStudySuggestions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Çalışma önerileri başarıyla oluşturuldu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının durumuna göre çalışma önerileri alma",
        "tags": [
          "AI Chat"
        ]
      }
    },
    "/ai-chat/explain-topic": {
      "post": {
        "operationId": "AiChatController_explainTopic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Konu açıklaması başarıyla oluşturuldu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Belirli bir konunun AI tarafından açıklanması",
        "tags": [
          "AI Chat"
        ]
      }
    },
    "/ai-chat/motivational-support": {
      "post": {
        "operationId": "AiChatController_getMotivationalSupport",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Motivasyon desteği başarıyla oluşturuldu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcıya motivasyon desteği sağlama",
        "tags": [
          "AI Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "🎓 GÜNÜBİRLİK DOZ API",
...
    "version": "2.0.0",
    "contact": {
      "name": "GÜNÜBİRLİK DOZ Geliştirici Ekibi",
      "url": "https://github.com/gunibirlikdoz",
      "email": "info@gunibirlikdoz.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT Bearer token ile kimlik doğrulama",
        "in": "header"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT access token (expires in 15 minutes)"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT refresh token (expires in 7 days)"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Refresh token to get new access token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-string",
            "description": "User unique identifier"
          },
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email address"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "role": {
            "type": "string",
            "example": "STUDENT",
            "description": "User role in the system",
            "enum": [
              "STUDENT",
              "TEACHER",
              "ADMIN",
              "PARENT"
            ]
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true,
            "description": "Whether user email is verified"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01T00:00:00.000Z",
            "description": "User birth date"
          },
          "age": {
            "type": "number",
            "example": 25,
            "description": "User age"
          },
          "gender": {
            "type": "string",
            "example": "MALE",
            "description": "User gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_TO_SAY"
            ]
          },
          "gradeLevel": {
            "type": "string",
            "example": "GRADE_10",
            "description": "User grade level",
            "enum": [
              "GRADE_1",
              "GRADE_2",
              "GRADE_3",
              "GRADE_4",
              "GRADE_5",
              "GRADE_6",
              "GRADE_7",
              "GRADE_8",
              "GRADE_9",
              "GRADE_10",
              "GRADE_11",
              "GRADE_12"
            ]
          },
          "onboardingStatus": {
            "type": "string",
            "example": "COMPLETED",
            "description": "User onboarding status",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "preferences": {
            "type": "object",
            "example": {
              "learningStyle": "VISUAL",
              "subjects": [
                "Math",
                "Science"
              ]
            },
            "description": "User preferences and learning settings"
          },
          "weakSubjects": {
            "example": [
              "Math",
              "Physics"
            ],
            "description": "Subjects where user needs improvement",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Account creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Last profile update date"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "onboardingStatus",
          "weakSubjects",
          "createdAt",
          "updatedAt"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "abc123def456",
            "description": "Email verification token"
          }
        },
        "required": [
          "token"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "oldpassword123",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "example": "newpassword123",
            "description": "New password"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "abc123def456",
            "description": "Password reset token"
          },
          "newPassword": {
            "type": "string",
            "example": "newpassword123",
            "description": "New password"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
    }
  },
...
}